+++ test/spec/unit/SocialSpec.js
@@ -73,7 +73,7 @@ describe("Social Unit Tests", function() {
 
     var user = new StackMob.User({ username: "testUser", password: "testUser" });
 
-    expectEndpoint(user, ['', '', '', false], 'loginWithGigya', 'POST', 'user/gigyaAccessToken')
+    expectEndpoint(user, ['', '', '', false], 'loginWithGigya', 'POST', 'user/gigyaAccessToken');
 
   });